requirements.txt
tensorflow>=2.12.0
gradio>=4.0.0
numpy
Pillow
